refactor(catalog): move CRUD/list/detail registrars into engine/entity_api (entity-engine phase C, part 2)
Relocate the generic HTTP-surface registrars out of catalog/api/ into the
domain-neutral engine, the headline of phase C. Wire URLs are byte-identical
(make codegen produces no frontend churn); the full backend suite is the oracle.

Moves:
- entity_list -> engine/entity_api/listing/ (paginated_list_response + q-fold)
- entity_detail_page -> engine/entity_api/detail/
- entity_crud splits across families: register_entity_create + the create
  primitives (entity_create) -> engine/entity_api/create/{registrar,persist};
  register_entity_delete_restore + soft_delete engine -> engine/entity_api/delete/
- generic query substrate -> engine/query/ (_facet_helpers -> facet_helpers,
  constants); per-entity facet configs stay domain-side
- generic delete wire schemas (BlockingReferrer/SoftDeleteBlocked/AlreadyDeleted/
  DeletePreviewBase/TaxonomyDeletePreview/DeleteResponse) -> delete/schemas.py;
  domain previews stay in catalog.api.schemas and subclass the base

Each operation family is the public API for its family; per-entity routers mount
the registrars by importing from the family subpackages. engine/__init__ stays
import-light (loaded during model setup) — no eager registrar imports.

Contracts:
- new "Engine internal layers" exhaustive contract (registrars on top, generic
  substrate at the base); the entity_api ⊄ export_api seam waits for Phase E
- AppBoundaries.md updated; the authz-enforcement test now filters the factory
  routes across both the create and delete registrar modules

No migration, no wire-URL change. lint-imports: 17 kept, 0 broken.
