+++ backend/apps/catalog/api/gameplay_features.py
@@ -23,12 +23,13 @@ from apps.media.selectors import serialize_uploaded_media
 from apps.provenance.helpers import claims_prefetch
 from apps.provenance.rate_limits import EDIT_RATE_LIMIT_SPEC, rate_limited
 
+from ..engine.entity_api.create import register_entity_create
+from ..engine.entity_api.delete import register_entity_delete_restore
+from ..engine.entity_api.listing import _apply_list_q
+from ..engine.query.constants import DEFAULT_PAGE_SIZE, NameAliasQuery, PageParam
 from ..models import GameplayFeature
 from ._counts import bulk_title_counts_via_models
-from .constants import DEFAULT_PAGE_SIZE, NameAliasQuery, PageParam
 from .edit_claims import plan_alias_claims, plan_parent_claims
-from .entity_crud import register_entity_create, register_entity_delete_restore
-from .entity_list import _apply_list_q
 from .rich_text import describe
 from .schemas import (
     EntityDetailSchema,
