+++ backend/apps/catalog/api/machine_models.py
@@ -48,6 +48,14 @@ from apps.provenance.schemas import (
     ChangeSetInputSchema,
 )
 
+from ..engine.entity_api.delete import (
+    SoftDeleteBlockedError,
+    count_entity_changesets,
+    execute_soft_delete,
+    plan_soft_delete,
+    serialize_blocking_referrer,
+)
+from ..engine.query.constants import DEFAULT_PAGE_SIZE
 from ..models import (
     Cabinet,
     Credit,
@@ -65,7 +73,6 @@ from ..models import (
     TechnologySubgeneration,
     Theme,
 )
-from .constants import DEFAULT_PAGE_SIZE
 from .edit_claims import (
     plan_abbreviation_claims,
     plan_credit_claims,
@@ -97,13 +104,6 @@ from .schemas import (
     ModelEditOptionsSchema,
     SoftDeleteBlockedSchema,
     TitleModelSchema,
-)
-from .soft_delete import (
-    SoftDeleteBlockedError,
-    count_entity_changesets,
-    execute_soft_delete,
-    plan_soft_delete,
-    serialize_blocking_referrer,
 )
 
 # ---------------------------------------------------------------------------
