+++ backend/apps/catalog/api/manufacturers.py
@@ -31,6 +31,9 @@ from ..cache import (
     manufacturers_facets_key,
     set_cached_response,
 )
+from ..engine.entity_api.create import register_entity_create
+from ..engine.entity_api.delete import register_entity_delete_restore
+from ..engine.query.constants import DEFAULT_PAGE_SIZE
 from ..models import (
     CorporateEntity,
     CorporateEntityLocation,
@@ -48,8 +51,6 @@ from ._manufacturer_facets import (
     query_count,
 )
 from ._typing import FacetOptionDict, HasModelCount
-from .constants import DEFAULT_PAGE_SIZE
-from .entity_crud import register_entity_create, register_entity_delete_restore
 from .helpers import (
     collect_titles,
     model_year_bounds,
