+++ backend/apps/catalog/api/people.py
@@ -43,17 +43,24 @@ from apps.provenance.rate_limits import (
 )
 from apps.provenance.schemas import ChangeSetInputSchema
 
-from ..models import Credit, MachineModel, Person
-from ._typing import HasCreditCount
-from .constants import NameAliasQuery, PageParam
-from .entity_create import (
+from ..engine.entity_api.create import (
     assert_name_available,
     assert_public_id_available,
     create_entity_with_claims,
     validate_name,
     validate_slug_format,
 )
-from .entity_list import _apply_list_q, paginated_list_response
+from ..engine.entity_api.delete import (
+    SoftDeleteBlockedError,
+    count_entity_changesets,
+    execute_soft_delete,
+    plan_soft_delete,
+    serialize_blocking_referrer,
+)
+from ..engine.entity_api.listing import _apply_list_q, paginated_list_response
+from ..engine.query.constants import NameAliasQuery, PageParam
+from ..models import Credit, MachineModel, Person
+from ._typing import HasCreditCount
 from .images import (
     extract_image_urls,
     fetch_model_media_map,
@@ -68,13 +75,6 @@ from .schemas import (
     PersonDeletePreviewSchema,
     PersonSoftDeleteBlockedSchema,
     RelatedTitleSchema,
-)
-from .soft_delete import (
-    SoftDeleteBlockedError,
-    count_entity_changesets,
-    execute_soft_delete,
-    plan_soft_delete,
-    serialize_blocking_referrer,
 )
