+++ backend/apps/catalog/api/series.py
@@ -22,11 +22,12 @@ from apps.provenance.helpers import claims_prefetch
 from apps.provenance.rate_limits import EDIT_RATE_LIMIT_SPEC, rate_limited
 from apps.provenance.schemas import RichTextSchema
 
+from ..engine.entity_api.create import register_entity_create
+from ..engine.entity_api.delete import register_entity_delete_restore
+from ..engine.entity_api.listing import paginated_list_response
+from ..engine.query.constants import NameQuery, PageParam
 from ..models import Credit, MachineModel, Series, Title
 from ._typing import HasTitleCount
-from .constants import NameQuery, PageParam
-from .entity_crud import register_entity_create, register_entity_delete_restore
-from .entity_list import paginated_list_response
 from .helpers import (
     serialize_credit,
     serialize_title_ref,
