+++ backend/apps/catalog/api/taxonomy.py
@@ -24,6 +24,10 @@ from apps.core.schemas import RateLimitErrorSchema, ValidationErrorSchema
 from apps.provenance.helpers import claims_prefetch
 from apps.provenance.rate_limits import EDIT_RATE_LIMIT_SPEC, rate_limited
 
+from ..engine.entity_api.create import register_entity_create
+from ..engine.entity_api.delete import register_entity_delete_restore
+from ..engine.entity_api.listing import paginated_list_response
+from ..engine.query.constants import NameAliasQuery, NameQuery, PageParam
 from ..models import (
     Cabinet,
     CatalogModel,
@@ -42,12 +46,6 @@ from ..models import (
 )
 from ._counts import bulk_title_counts_via_models
 from ._typing import HasTitleCount
-from .constants import NameAliasQuery, NameQuery, PageParam
-from .entity_crud import (
-    register_entity_create,
-    register_entity_delete_restore,
-)
-from .entity_list import paginated_list_response
 from .helpers import serialize_title_machine
 from .images import extract_image_urls, fetch_model_media_map
 from .people import PersonCardSchema
