+++ backend/apps/catalog/api/themes.py
@@ -21,12 +21,13 @@ from apps.core.schemas import RateLimitErrorSchema, ValidationErrorSchema
 from apps.provenance.helpers import claims_prefetch
 from apps.provenance.rate_limits import EDIT_RATE_LIMIT_SPEC, rate_limited
 
+from ..engine.entity_api.create import register_entity_create
+from ..engine.entity_api.delete import register_entity_delete_restore
+from ..engine.entity_api.listing import _apply_list_q
+from ..engine.query.constants import DEFAULT_PAGE_SIZE, NameAliasQuery, PageParam
 from ..models import MachineModel, Theme
 from ._counts import bulk_title_counts_via_models
-from .constants import DEFAULT_PAGE_SIZE, NameAliasQuery, PageParam
 from .edit_claims import plan_alias_claims, plan_parent_claims
-from .entity_crud import register_entity_create, register_entity_delete_restore
-from .entity_list import _apply_list_q
 from .helpers import serialize_title_machine
 from .images import fetch_model_media_map
 from .rich_text import describe
