+++ docs/AppBoundaries.md
@@ -54,7 +54,7 @@ Within an app, a few internal boundaries are enforced (named contracts in `pypro
 - **The api/schema layer is a leaf.** Domain code (models, services, …) must not import back up into `api/`. Both catalog and core express this as an _exhaustive_ layered contract over the app's whole acyclic internal stack (`Catalog internal layers`, `Core internal layers`) — a new submodule left unplaced fails the build. The remaining apps are single `api.py` modules, leaves by construction.
 - **Domain models do not import the api/schema layer** — a model reaching up into HTTP serialization is a dependency inversion.
 - **The generic resolver core stays catalog-agnostic.** `catalog/resolve/{_entities,_helpers,_claim_values}` drive entirely off claim-field introspection and name no concrete model; the contract forbids them from importing the catalog domain so that property can't silently regress.
-- **The domain-neutral engine stays catalog-agnostic.** `catalog/engine/` — the generic HTTP-surface substrate plus the generic bases the domain mounts and subclasses (`AliasModel`, the wire schema bases, `naming`, the model-registry walk) — sits at the bottom of the `Catalog internal layers` stack (concrete alias models in `models/` subclass `engine.models.AliasModel`, so `models → engine`). `Engine does not import the catalog domain` forbids the reverse edge, so the engine names no concrete model and stays liftable to its own app by a `git mv`.
+- **The domain-neutral engine stays catalog-agnostic.** `catalog/engine/` — the generic HTTP-surface registrars (`entity_api/{create,detail,listing,delete}`), the query fold (`query/`) and the generic bases the domain mounts and subclasses (`AliasModel`, the wire schema bases, `naming`, the model-registry walk) — sits at the bottom of the `Catalog internal layers` stack (concrete alias models in `models/` subclass `engine.models.AliasModel`, so `models → engine`). `Engine does not import the catalog domain` forbids the reverse edge, so the engine names no concrete model and stays liftable to its own app by a `git mv`. Its own internals are exhaustively layered by `Engine internal layers` (registrars on top, the generic substrate at the base).
 - **The patch front end does not depend on the ingest back end** — `claim_ingest.patches` lowers YAML to an `IngestPlan` that `claim_ingest.apply` executes; new patch behavior belongs against that IR, not as a fork of the back end.
 - **Production code does not import test factories** — `test_factories` is test-only scaffolding and must not enter the runtime path.
